+++ internal/client/dialer.go
@@ -6,28 +6,30 @@ import (
 	"fmt"
 	"net"
 
+	"github.com/ameshkov/gocurl/internal/client/connectto"
+	"github.com/ameshkov/gocurl/internal/client/dialer"
+	"github.com/ameshkov/gocurl/internal/client/ech"
+	"github.com/ameshkov/gocurl/internal/client/proxy"
+	"github.com/ameshkov/gocurl/internal/client/splittls"
 	"github.com/ameshkov/gocurl/internal/config"
 	"github.com/ameshkov/gocurl/internal/output"
 	"github.com/ameshkov/gocurl/internal/resolve"
 	"github.com/quic-go/quic-go"
-	"golang.org/x/net/proxy"
 )
 
-type dialFunc func(network, addr string) (net.Conn, error)
-
-// dialer is a structure that implements additional logic on top of the
+// clientDialer is a structure that implements additional logic on top of the
 // regular dial depending on the configuration. It can dial over a proxy,
 // apply --connect-to logic or split TLS client hello when required.
-type dialer struct {
+type clientDialer struct {
 	cfg       *config.Config
 	out       *output.Output
 	tlsConfig *tls.Config
 	resolver  *resolve.Resolver
-	dial      dialFunc
+	dial      dialer.DialFunc
 }
 
-// newDialer creates a new instance of the dialer.
-func newDialer(cfg *config.Config, out *output.Output) (d *dialer, err error) {
+// newDialer creates a new instance of the clientDialer.
+func newDialer(cfg *config.Config, out *output.Output) (d *clientDialer, err error) {
 	resolver, err := resolve.NewResolver(cfg, out)
 	if err != nil {
 		return nil, err
@@ -38,7 +40,7 @@ func newDialer(cfg *config.Config, out *output.Output) (d *dialer, err error) {
 		return nil, err
 	}
 
-	return &dialer{
+	return &clientDialer{
 		cfg:       cfg,
 		out:       out,
 		tlsConfig: createTLSConfig(cfg),
@@ -47,11 +49,8 @@ func newDialer(cfg *config.Config, out *output.Output) (d *dialer, err error) {
 	}, nil
 }
 
-// type check
-var _ proxy.ContextDialer = (*dialer)(nil)
-
 // DialTLSContext establishes a new TLS connection to the specified address.
-func (d *dialer) DialTLSContext(_ context.Context, network, addr string) (c net.Conn, err error) {
+func (d *clientDialer) DialTLSContext(_ context.Context, network, addr string) (c net.Conn, err error) {
 	d.out.Debug("Connecting to %s over TLS", addr)
 
 	conn, err := d.dial(network, addr)
@@ -59,25 +58,44 @@ func (d *dialer) DialTLSContext(_ context.Context, network, addr string) (c net.
 		return nil, err
 	}
 
-	tlsConn := tls.Client(conn, d.tlsConfig)
-	err = tlsConn.Handshake()
-	if err != nil {
-		return nil, err
+	if d.cfg.ECH {
+		return d.handshakeECH(conn)
 	}
 
-	return tlsConn, nil
+	return d.handshakeTLS(conn)
 }
 
-// DialContext implements proxy.ContextDialer for *dialer.
-func (d *dialer) DialContext(_ context.Context, network, addr string) (c net.Conn, err error) {
+// DialContext implements proxy.ContextDialer for *clientDialer.
+func (d *clientDialer) DialContext(_ context.Context, network, addr string) (c net.Conn, err error) {
 	d.out.Debug("Connecting to %s", addr)
 
 	return d.dial(network, addr)
 }
 
+// handshakeTLS attempts to establish a TLS connection.
+func (d *clientDialer) handshakeTLS(conn net.Conn) (tlsConn net.Conn, err error) {
+	tlsClient := tls.Client(conn, d.tlsConfig)
+	err = tlsClient.Handshake()
+	if err != nil {
+		return nil, err
+	}
+
+	return tlsClient, nil
+}
+
+// handshakeECH attempts to establish a ECH-enabled TLS connection.
+func (d *clientDialer) handshakeECH(conn net.Conn) (tlsConn net.Conn, err error) {
+	echConfigs, err := d.resolver.LookupECHConfigs(d.tlsConfig.ServerName)
+	if err != nil {
+		return nil, err
+	}
+
+	return ech.HandshakeECH(conn, echConfigs, d.tlsConfig, d.out)
+}
+
 // DialQUIC establishes a new QUIC connection and is supposed to be used by
 // http3.RoundTripper.
-func (d *dialer) DialQUIC(
+func (d *clientDialer) DialQUIC(
 	ctx context.Context,
 	addr string,
 	_ *tls.Config,
@@ -101,93 +119,35 @@ func (d *dialer) DialQUIC(
 	return quic.DialEarly(ctx, uConn, udpAddr, d.tlsConfig, cfg)
 }
 
-// udpConn is a wrapper over a pre-connected net.PacketConn that overrides
-// WriteTo and ReadFrom methods to make it work.
-type udpConn struct {
-	net.Conn
-}
-
-// ReadFrom implements net.PacketConn for udpConn.
-func (u *udpConn) ReadFrom(b []byte) (n int, addr net.Addr, err error) {
-	n, err = u.Read(b)
-
-	return n, u.RemoteAddr(), err
-}
-
-// WriteTo implements net.PacketConn for udpConn.
-func (u *udpConn) WriteTo(b []byte, _ net.Addr) (n int, err error) {
-	return u.Write(b)
-}
-
-// type check
-var _ net.PacketConn = (*udpConn)(nil)
-
-// directDialer provides the base dialFunc implementation.
-type directDialer struct {
-	out      *output.Output
-	resolver *resolve.Resolver
-}
-
-// type check
-var _ proxy.Dialer = (*directDialer)(nil)
-
-// Dial implements proxy.Dialer for *directDialer.
-func (d *directDialer) Dial(network, addr string) (conn net.Conn, err error) {
-	d.out.Debug("Connecting to %s://%s", network, addr)
-
-	host, port, err := net.SplitHostPort(addr)
-	if err != nil {
-		return nil, err
-	}
-
-	ipAddrs, err := d.resolver.LookupHost(host)
-	if err != nil {
-		return nil, err
-	}
-
-	ipAddr := ipAddrs[0]
-
-	conn, err = net.Dial(network, net.JoinHostPort(ipAddr.String(), port))
-	if err != nil {
-		return nil, err
-	}
-
-	if _, ok := conn.(net.PacketConn); ok {
-		return &udpConn{Conn: conn}, nil
-	}
-
-	return conn, nil
-}
-
 // createDialFunc creates dialFunc that implements all the logic configured by
 // cfg.
 func createDialFunc(
 	resolver *resolve.Resolver,
 	cfg *config.Config,
 	out *output.Output,
-) (dial dialFunc, err error) {
-	d := &directDialer{
-		out:      out,
-		resolver: resolver,
-	}
+) (dial dialer.DialFunc, err error) {
+	d := dialer.NewDirect(resolver, out)
 	dial = d.Dial
 
 	if cfg.ProxyURL != nil {
-		dial, err = createProxyDialFunc(cfg.ProxyURL, dial, out)
+		var proxyDialer dialer.Dialer
+		proxyDialer, err = proxy.NewProxyDialer(cfg.ProxyURL, dial, out)
 		if err != nil {
 			return nil, err
 		}
+
+		dial = proxyDialer.Dial
 	}
 
 	if len(cfg.ConnectTo) > 0 {
-		dial, err = createConnectToDialFunc(cfg.ConnectTo, dial, out)
+		dial, err = connectto.CreateDialFunc(cfg.ConnectTo, dial, out)
 		if err != nil {
 			return nil, err
 		}
 	}
 
 	if cfg.TLSSplitChunkSize > 0 {
-		dial = createTLSSplitDialFunc(cfg.TLSSplitChunkSize, cfg.TLSSplitDelay, dial, out)
+		dial = splittls.CreateDialFunc(cfg.TLSSplitChunkSize, cfg.TLSSplitDelay, dial, out)
 	}
 
 	return dial, nil