+++ internal/ech/ech.go
@@ -1,9 +1,0 @@
-package ech
-
-func dial() {
-	//tlsConf := tls.Config{
-	//	ECHEnabled: true,
-	//}
-	//
-	//client := http.Client{}
-}